backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-telegram-bot==21.6
httpx==0.27.2
redis[asyncio]==5.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
aiofiles==24.1.0
cryptography==43.0.3
pyotp==2.9.0
qrcode[pil]==8.0
aiosmtplib==3.0.1
jinja2==3.1.4
openpyxl==3.1.5
reportlab==4.2.5
pillow==10.4.0
python-dotenv==1.0.1
slowapi==0.1.9
